+++ pyriemann/clustering.py
@@ -8,14 +8,17 @@ from scipy.stats import norm, chi2
 import sklearn
 from sklearn.base import BaseEstimator, TransformerMixin, clone
 from sklearn.cluster import KMeans as sklearnKMeans
+from sklearn.utils.validation import check_random_state
 
 from ._base import SpdClassifMixin, SpdClustMixin, SpdTransfMixin
 from .classification import MDM
-from .utils.distance import distance, pairwise_distance
+from .datasets import sample_gaussian_spd
+from .utils.distance import distance, pairwise_distance, distance_mahalanobis
 from .utils.mean import mean_covariance
 from .utils.geodesic import geodesic
 from .utils.tangentspace import exp_map, log_map
-from .utils.utils import check_metric, check_function
+from .utils.utils import check_metric, check_function, check_weights
+from .utils.tangentspace import tangent_space
 
 
 def _init_centroids(X, n_clusters, init, random_state, x_squared_norms):
@@ -341,6 +344,9 @@ class KmeansPerClassTransform(SpdTransfMixin, BaseEstimator):
         return mdm._predict_distances(X)
 
 
+###############################################################################
+
+
 @np.vectorize
 def kernel_normal(x):
     return np.exp(- x ** 2)
@@ -448,7 +454,9 @@ class MeanShift(SpdClustMixin, BaseEstimator):
             The MeanShift instance.
         """
         self._kernel_fun = check_function(self.kernel, ker_clust_functions)
-        self.metric_map, self.metric_dist = check_metric(self.metric)
+        self._metric_map, self._metric_dist = check_metric(
+            self.metric, ["map", "dist"]
+        )
         if self.bandwidth is None:
             self._bandwidth = self._estimate_bandwidth(X, quantile=0.3)
         self._bandwidth2 = self._bandwidth ** 2
@@ -465,7 +473,7 @@ class MeanShift(SpdClustMixin, BaseEstimator):
         return self
 
     def _estimate_bandwidth(self, X, quantile):
-        dist = pairwise_distance(X, None, metric=self.metric_dist)
+        dist = pairwise_distance(X, None, metric=self._metric_dist)
         dist = np.triu(dist, 1)
         dist_sorted = np.sort(dist[dist > 0])
         bandwidth = dist_sorted[floor(quantile * len(dist_sorted))]
@@ -474,11 +482,11 @@ class MeanShift(SpdClustMixin, BaseEstimator):
 
     def _seek_mode(self, X, mean):
         for _ in range(self.max_iter):
-            T = log_map(X, mean, metric=self.metric_map)
-            dist2 = distance(X, mean, metric=self.metric_dist, squared=True)
+            T = log_map(X, mean, metric=self._metric_map)
+            dist2 = distance(X, mean, metric=self._metric_dist, squared=True)
             weights = self._kernel_fun(dist2[:, 0] / self._bandwidth2)
             meanshift = np.einsum("a,abc->bc", weights, T) / np.sum(weights)
-            mean = exp_map(meanshift, mean, metric=self.metric_map)
+            mean = exp_map(meanshift, mean, metric=self._metric_map)
             if np.linalg.norm(meanshift) <= self.tol:
                 break
         else:
@@ -489,7 +497,7 @@ class MeanShift(SpdClustMixin, BaseEstimator):
     def _fuse_mode(self, in_modes):
         out_modes = in_modes.copy()
         in_modes = np.stack(in_modes, axis=0)
-        dist = pairwise_distance(in_modes, None, metric=self.metric_dist)
+        dist = pairwise_distance(in_modes, None, metric=self._metric_dist)
         np.fill_diagonal(dist, self._bandwidth + 1)
         for i in range(dist.shape[0] - 1, -1, -1):
             if np.min(dist[i]) < self._bandwidth:
@@ -518,11 +526,422 @@ class MeanShift(SpdClustMixin, BaseEstimator):
             Prediction for each matrix according to the closest mode.
         """
         dist = Parallel(n_jobs=self.n_jobs)(
-            delayed(distance)(X, mode, self.metric_dist)
+            delayed(distance)(X, mode, self._metric_dist)
             for mode in self.modes_
         )
         dist = np.concatenate(dist, axis=1)
         return dist.argmin(axis=1)
+
+
+###############################################################################
+
+
+class Gaussian():
+    """Gaussian model.
+
+    Gaussian model for Riemannian manifold of SPD matrices,
+    defined with a mean in manifold and a covariance in tangent space [1]_.
+
+    Parameters
+    ----------
+    n : integer
+        Dimension of the matrices.
+    mu : ndarray, shape (n, n)
+        Mean of the Gaussian, in manifold.
+    sigma : None | ndarray, shape (n * (n + 1) / 2, n * (n + 1) / 2), \
+            default=None
+        Covariance of the Gaussian, in tangent space.
+        If None, it uses identity matrix.
+    metric : string | dict, default="riemann"
+        Metric used for mean update (for the list of supported metrics,
+        see :func:`pyriemann.utils.mean.mean_covariance`) and
+        for tangent space map
+        (see :func:`pyriemann.utils.tangent_space.tangent_space`).
+        The metric can be a dict with two keys, "mean" and "map"
+        in order to pass different metrics.
+
+    Notes
+    -----
+    .. versionadded:: 0.11
+
+    References
+    ----------
+    .. [1] `Intrinsic statistics on Riemannian manifolds: Basic tools for
+        geometric measurements
+        <https://www.cis.jhu.edu/~tingli/App_of_Lie_group/Intrinsic%20Statistics%20on%20Riemannian%20Manifolds.pdf>`_
+        X. Pennec. Journal of Mathematical Imaging and Vision, 2006
+    """  # noqa
+    def __init__(self, n, mu, sigma=None, metric="riemann"):
+        self.n = n
+        self.mu = mu
+        if sigma is None:
+            sigma = np.eye(n * (n + 1) // 2)
+        self.sigma = sigma
+        self.metric = metric
+        self._metric_mean, self._metric_map = check_metric(
+            metric, ["mean", "map"]
+        )
+
+    def pdf(self, X, *, reg=1e-16, use_pi=True):
+        """Compute approximate probability density function (pdf) of matrices.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n, n)
+            Set of SPD matrices.
+        reg : float, default=1e-16
+            Regularization parameter for pdf normalization term.
+        use_pi : bool, default=True
+            If true, use (2 pi)^n to compute the full denominator.
+            If false, do not use (2 pi)^n, because will be simplified with
+            upcoming normalizations.
+
+        Returns
+        -------
+        pdf : ndarray, shape (n_matrices,)
+            Probability density function of each matrix.
+        """
+        TangVec = tangent_space(X, self.mu, metric=self._metric_map)
+        dist = distance_mahalanobis(TangVec.T, self.sigma, squared=True)
+        num = np.exp(-0.5 * dist)
+        det = np.linalg.det(self.sigma)
+        if use_pi:
+            denom = np.sqrt(((2 * np.pi) ** self.n) * det)
+        else:
+            denom = np.sqrt(det)
+        return num / (denom + reg)
+
+    def update_mean(self, X, sample_weight):
+        """Update mean in manifold.
+
+        Compute weighted mean of matrices, initialized on previous mean.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n, n)
+            Set of SPD matrices.
+        sample_weight : ndarray, shape (n_matrices,)
+            Weights for each matrix.
+        """
+        self.mu = mean_covariance(
+            X,
+            metric=self._metric_mean,
+            sample_weight=sample_weight,
+            init=self.mu,
+        )
+
+    def update_covariance(self, X, sample_weight):
+        """Update covariance in tangent space.
+
+        Compute weighted covariance of tangent vectors.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n, n)
+            Set of SPD matrices.
+        sample_weight : ndarray, shape (n_matrices,)
+            Weights for each matrix.
+        """
+        TangVec = tangent_space(X, self.mu, metric=self._metric_map)
+        sigma = TangVec.T @ (sample_weight[:, np.newaxis] * TangVec)
+        self.sigma = sigma / sample_weight.sum()
+
+
+class GaussianMixture(SpdClustMixin, BaseEstimator):
+    """Gaussian mixture model.
+
+    Representation of a Gaussian mixture model (GMM) probability distribution
+    for SPD matrices by expectation-maximization (EM) algorithm [1]_.
+
+    Parameters
+    ----------
+    n_components : integer, default=1
+        Number of mixture components.
+    metric : string | dict, default="riemann"
+        Metric used for mean update (for the list of supported metrics,
+        see :func:`pyriemann.utils.mean.mean_covariance`) and
+        for tangent space map
+        (see :func:`pyriemann.utils.tangent_space.tangent_space`).
+        The metric can be a dict with two keys, "mean" and "map"
+        in order to pass different metrics.
+    weights_init : None | ndarray, shape (n_components,), defaut=None
+        Initial weights. If None, it uses equal weights.
+    means_init : None | ndarray, shape (n_components,), defaut=None
+        Initial means of Gaussians. If None, it randomly selects training
+        matrices.
+    tol : float, default=1e-5
+        Tolerance to stop the EM algorithm.
+    maxiter : int, default=100
+        Maximum number of iterations of EM algorithm.
+    random_state : None | integer | np.RandomState, default=None
+        The generator used to initialize the Gaussian models. If an integer is
+        given, it fixes the seed. Defaults to the global numpy random
+        number generator.
+    verbose : bool, default=False
+        Verbose flag.
+
+    Attributes
+    ----------
+    weights_ : ndarray, shape (n_components,)
+        Weight of each mixture component.
+    means_ : ndarray, shape (n_components, n_channels, n_channels)
+        Mean of each mixture component.
+    covariances_ : ndarray, shape (n_components, n_ts, n_ts)
+        Covariance of each mixture component.
+
+    Notes
+    -----
+    .. versionadded:: 0.11
+
+    References
+    ----------
+    .. [1] `Gaussian mixture regression on symmetric positive definite matrices
+        manifolds: Application to wrist motion estimation with sEMG
+        <https://calinon.ch/papers/Jaquier-IROS2017.pdf>`_
+        N. Jacquier & S. Calinon. IEEE IROS, 2017
+    """
+    def __init__(
+        self,
+        n_components=1,
+        metric="riemann",
+        weights_init=None,
+        means_init=None,
+        tol=1e-5,
+        maxiter=100,
+        random_state=None,
+        verbose=False,
+    ):
+        """Init."""
+        self.n_components = n_components
+        self.metric = metric
+        self.weights_init = weights_init
+        self.means_init = means_init
+        self.tol = tol
+        self.maxiter = maxiter
+        self.random_state = random_state
+        self.verbose = verbose
+
+    @property
+    def means_(self):
+        return np.stack([component.mu for component in self._components])
+
+    @property
+    def covariances_(self):
+        return np.stack([component.sigma for component in self._components])
+
+    def _get_wlik(self, X, use_pi=True):
+        """Compute weighted likelihoods.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+        use_pi : bool, default=True
+            If true, use (2 pi)^n to compute the full denominator of pdf.
+            If false, do not use (2 pi)^n, because will be simplified with
+            upcoming normalizations.
+
+        Returns
+        -------
+        wlik : ndarray, shape (n_matrices, n_components)
+            Weighted likelihood of each matrix given component.
+        """
+        wlik = np.zeros((X.shape[0], self.n_components))
+        for k in range(self.n_components):
+            lik = self._components[k].pdf(X, use_pi=use_pi)
+            wlik[:, k] = self.weights_[k] * lik
+        return wlik
+
+    def _get_proba(self, X, reg=1e-16):
+        """Compute posterior probabilities.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+        reg : float, default=1e-16
+            Regularization parameter for probabilities normalization.
+
+        Returns
+        -------
+        prob : ndarray, shape (n_matrices, n_components)
+            Posterior probability of each component given matrix.
+        """
+        num = self._get_wlik(X, use_pi=False)
+        prob = num / (np.sum(num, axis=1, keepdims=True) + reg)
+        return prob
+
+    def _log(self, X):
+        """Log after clip."""
+        return np.log(np.clip(X, a_min=1e-10, a_max=None))
+
+    def fit(self, X, y=None):
+        """Fit the mixture with EM.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+        y : None
+            Not used, here for compatibility with sklearn API.
+
+        Returns
+        -------
+        self : GaussianMixture instance
+            The GaussianMixture instance.
+        """
+        n_matrices, n_channels, _ = X.shape
+        if (n_channels * (n_channels + 1) // 2 > n_matrices):
+            raise ValueError("Not enough matrices for training GMM.")
+
+        # initialization
+        self.random_state = check_random_state(self.random_state)
+
+        if isinstance(self.means_init, np.ndarray) and self.means_init.shape \
+                == (self.n_components, n_channels, n_channels):
+            means_init = self.means_init
+        else:
+            inds = self.random_state.randint(
+                n_matrices,
+                size=(self.n_components,)
+            )
+            means_init = X[inds]
+
+        self._components = []
+        for k in range(self.n_components):
+            self._components.append(
+                Gaussian(
+                    n_channels,
+                    mu=means_init[k],
+                    sigma=None,
+                    metric=self.metric,
+                )
+            )
+
+        self.weights_ = check_weights(self.weights_init, self.n_components)
+
+        # expectation-maximization
+        crit = 0
+        for _ in range(self.maxiter):
+            # e-step
+            prob = self._get_proba(X)
+
+            # m-step
+            self.weights_ = np.sum(prob, axis=0) / n_matrices
+            # re-normalization (necessary because of approx Gaussian pdf?)
+            self.weights_ = self.weights_ / self.weights_.sum()
+            for k in range(self.n_components):
+                self._components[k].update_mean(X, prob[:, k])
+                self._components[k].update_covariance(X, prob[:, k])
+
+            # check convergence
+            crit_new = -np.sum(self._log(np.sum(self._get_wlik(X), axis=1)))
+            if self.verbose:
+                print(f"neg log-likelihood = {crit_new}")
+            if abs(crit - crit_new) < self.tol:
+                break
+            crit = crit_new
+        else:
+            warnings.warn("EM convergence not reached")
+
+        return self
+
+    def predict_proba(self, X):
+        """Predict probabilities.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+
+        Returns
+        -------
+        prob : ndarray, shape (n_matrices, n_components)
+            Probabilities for each component.
+        """
+        return self._get_proba(X)
+
+    def predict(self, X):
+        """Get the predictions.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+
+        Returns
+        -------
+        pred : ndarray of int, shape (n_matrices,)
+            Predictions for each matrix.
+        """
+        prob = self._get_proba(X)
+        return np.argmax(prob, axis=1)
+
+    def score(self, X, y=None):
+        """Compute the average log-likelihood of the given matrices.
+
+        Parameters
+        ----------
+        X : ndarray, shape (n_matrices, n_channels, n_channels)
+            Set of SPD matrices.
+        y : None
+            Not used, here for compatibility with sklearn API.
+
+        Returns
+        -------
+        score : float
+            Log-likelihood of matrices under the Gaussian mixture model.
+        """
+        lik = np.sum(self._get_wlik(X), axis=1)
+        return np.mean(self._log(lik))
+
+    def sample(self, n_matrices=1):
+        """Generate random matrices from the fitted Gaussian distribution.
+
+        Warning: GMM is calibrated using the Gaussian model [1]_,
+        while this sampling uses the wrapped Gaussian model [2]_.
+
+        Parameters
+        ----------
+        n_matrices : int, default=1
+            Number of matrices to generate.
+
+        Returns
+        -------
+        X : array, shape (n_matrices, n_channels, n_channels)
+            Randomly generated matrices.
+        y : array, shape (n_matrices,)
+            Component labels.
+
+        References
+        ----------
+        .. [1] `Intrinsic statistics on Riemannian manifolds: Basic tools for
+            geometric measurements
+            <https://www.cis.jhu.edu/~tingli/App_of_Lie_group/Intrinsic%20Statistics%20on%20Riemannian%20Manifolds.pdf>`_
+            X. Pennec. Journal of Mathematical Imaging and Vision, 2006
+        .. [2] `Wrapped gaussian on the manifold of symmetric positive
+            definite matrices
+            <https://openreview.net/pdf?id=EhStXG4dCS>`_
+            T. de Surrel, F. Lotte, S. Chevallier, and F. Yger. ICML, 2025
+        """  # noqa
+        y = self.random_state.randint(self.n_components, size=(n_matrices,))
+
+        means, covariances = self.means_, self.covariances_
+        n_channels = means.shape[-1]
+
+        X = np.zeros((n_matrices, means.shape[-1], n_channels))
+        for i in np.unique(y):
+            X[y == i] = sample_gaussian_spd(
+                np.count_nonzero(y == i),
+                mean=means[i],
+                sigma=covariances[i],
+                random_state=self.random_state
+            )
+
+        return X, y
+
+
+###############################################################################
 
 
 class Potato(TransformerMixin, SpdClassifMixin, BaseEstimator):
@@ -1100,7 +1519,7 @@ class PotatoField(TransformerMixin, SpdClassifMixin, BaseEstimator):
         for i in range(self.n_potatoes):
             _check_n_matrices(X[i], n_matrices)
             p[i] = self._potatoes[i].predict_proba(X[i])
-        p[p < 1e-10] = 1e-10  # avoid trouble with log
+        p = np.clip(p, a_min=1e-10, a_max=1)  # avoid trouble with log
         q = - 2 * np.sum(np.log(p), axis=0)
         proba = self._get_proba(q)
         return proba