+++ tests/test_clustering.py
@@ -1,23 +1,37 @@
 import numpy as np
 from numpy.testing import assert_array_equal
 import pytest
+from pytest import approx
 
 from pyriemann.clustering import (
     Kmeans,
     KmeansPerClassTransform,
     MeanShift,
+    Gaussian,
+    GaussianMixture,
     Potato,
     PotatoField,
 )
+from pyriemann.utils.tangentspace import tangent_space
 
-clusts = [Kmeans, KmeansPerClassTransform, MeanShift, Potato, PotatoField]
+clusts = [
+    Kmeans,
+    KmeansPerClassTransform,
+    MeanShift,
+    GaussianMixture,
+    Potato,
+    PotatoField,
+]
 
 
 @pytest.mark.parametrize("kind", ["spd", "hpd"])
 @pytest.mark.parametrize("clust", clusts)
 def test_clustering_two_clusters(kind, clust,
                                  get_mats, get_labels, get_weights):
-    n_clusters, n_matrices, n_channels = 2, 6, 4
+    if kind == "hpd" and clust in [GaussianMixture]:
+        pytest.skip()
+
+    n_clusters, n_matrices, n_channels = 2, 40, 3
     X = get_mats(n_matrices, n_channels, kind)
     weights = get_weights(n_matrices)
 
@@ -44,6 +58,13 @@ def test_clustering_two_clusters(kind, clust,
         clt_fit(clust, X, n_clusters, None)
         clt_predict(clust, X)
         clt_fitpredict(clust, X)
+
+    if clust is GaussianMixture:
+        clt_fit(clust, X, n_clusters, None)
+        clt_predict(clust, X)
+        clt_fitpredict(clust, X)
+        clt_predict_proba(clust, X)
+        clt_score(clust, X)
 
     if clust is Potato:
         clt_fit(clust, X, n_clusters, None)
@@ -74,7 +95,10 @@ def test_clustering_two_clusters(kind, clust,
 @pytest.mark.parametrize("kind", ["spd", "hpd"])
 @pytest.mark.parametrize("clust", clusts)
 def test_clustering_three_clusters(kind, clust, get_mats, get_labels):
-    n_clusters, n_matrices, n_channels = 3, 6, 2
+    if kind == "hpd" and clust in [GaussianMixture]:
+        pytest.skip()
+
+    n_clusters, n_matrices, n_channels = 3, 30, 2
     X = get_mats(n_matrices, n_channels, kind)
 
     if clust is Kmeans:
@@ -96,6 +120,13 @@ def test_clustering_three_clusters(kind, clust, get_mats, get_labels):
         clt_fit(clust, X, n_clusters, None)
         clt_predict(clust, X)
 
+    if clust is GaussianMixture:
+        clt_fit(clust, X, n_clusters, None)
+        clt_predict(clust, X)
+        clt_fitpredict(clust, X)
+        clt_predict_proba(clust, X)
+        clt_score(clust, X)
+
 
 def clt_fit(clust, X, n_clusters, labels):
     n_matrices, n_channels, _ = X.shape
@@ -109,14 +140,24 @@ def clt_fit(clust, X, n_clusters, labels):
         assert len(clt.mdm_.covmeans_) <= n_clusters
         assert clt.mdm_.covmeans_.shape[1:] == (n_channels, n_channels)
         assert_array_equal(clt.mdm_.covmeans_, clt.centroids())
+        return
     if clust is KmeansPerClassTransform:
         assert clt.classes_.shape == (n_classes,)
         assert len(clt.covmeans_) <= n_clusters * n_classes
         assert clt.covmeans_.shape[1:] == (n_channels, n_channels)
+        return
     if clust is MeanShift:
         assert clt.labels_.shape == (n_matrices,)
         assert len(clt.modes_) >= 1
         assert clt.modes_.shape[1:] == (n_channels, n_channels)
+        return
+    if clust is GaussianMixture:
+        assert clt.weights_.shape == (clt.n_components,)
+        assert clt.weights_.sum() == approx(1)
+        assert clt.means_.shape == (clt.n_components, n_channels, n_channels)
+        n_ts = n_channels * (n_channels + 1) // 2
+        assert clt.covariances_.shape == (clt.n_components, n_ts, n_ts)
+        return
     if clust is Potato:
         assert clt.covmean_.shape == (n_channels, n_channels)
 
@@ -194,6 +235,8 @@ def clt_fitpredict(clust, X, n_clusters=None):
     if hasattr(clt, "random_state"):
         clt.set_params(**{"random_state": 42})
     pred = clt.fit(X).predict(X)
+    if hasattr(clt, "random_state"):
+        clt.set_params(**{"random_state": 42})
     pred2 = clt.fit_predict(X)
     assert_array_equal(pred, pred2)
 
@@ -206,8 +249,10 @@ def clt_predict_proba(clust, X, n=None):
         n_matrices = len(X[0])
         clt = clust(n_potatoes=n)
     clt.fit(X)
-    proba = clt.predict_proba(X)
-    assert proba.shape == (n_matrices,)
+    prob = clt.predict_proba(X)
+    assert prob.shape[0] == n_matrices
+    if prob.ndim > 1:
+        assert prob.sum(axis=-1) == approx(np.ones(n_matrices))
 
 
 def clt_partial_fit(clust, X, n=None):
@@ -264,6 +309,12 @@ def clt_fittransform_per_class(clust, X, n_clusters, y):
     assert_array_equal(Xt, Xt2)
 
 
+def clt_score(clust, X, y=None):
+    clt = clust()
+    score = clt.fit(X, y).score(X, y)
+    assert isinstance(score, float)
+
+
 @pytest.mark.parametrize("clust", [Kmeans, KmeansPerClassTransform])
 @pytest.mark.parametrize("init", ["random", "ndarray"])
 @pytest.mark.parametrize("n_init", [1, 5])
@@ -316,6 +367,54 @@ def test_meanshift(kernel, metric, get_mats, get_labels):
         metric=metric,
     )
     clt.fit(X)
+
+
+def test_gaussian(get_mats, get_weights):
+    n_matrices, n = 13, 3
+    X = get_mats(n_matrices, n, "spd")
+    weights = get_weights(n_matrices)
+
+    gm = Gaussian(n=n, mu=X[0], metric="riemann")
+
+    gm.update_mean(X, weights)
+    assert gm.mu.shape == (n, n)
+
+    gm.update_covariance(X, weights)
+    n_ts = n * (n + 1) // 2
+    assert gm.sigma.shape == (n_ts, n_ts)
+
+    pdf = gm.pdf(X)
+    assert pdf.shape == (n_matrices,)
+
+    tv = tangent_space(X, gm.mu, metric="riemann")[0]
+    dist = tv.T @ np.linalg.inv(gm.sigma) @ tv
+    num = np.exp(-0.5 * dist)
+    denom = np.sqrt(((2 * np.pi) ** n) * np.linalg.det(gm.sigma))
+    pdf_ = num / (denom + 1e-16)
+    assert pdf[0] == approx(pdf_)
+
+
+@pytest.mark.parametrize("n_components", [2, 4])
+def test_gmm(n_components, get_mats, get_weights):
+    n_matrices, n_channels = 50, 2
+    X = get_mats(n_matrices, n_channels, "spd")
+    means_init = get_mats(n_components, n_channels, "spd")
+    weights_init = get_weights(n_components)
+
+    gmm = GaussianMixture(
+        n_components=n_components,
+        means_init=means_init,
+        weights_init=weights_init,
+    )
+    gmm.fit(X)
+
+    n_sampled_matrices = 20
+    X, y = gmm.sample(n_sampled_matrices)
+    assert X.shape == (n_sampled_matrices, n_channels, n_channels)
+    assert y.shape == (n_sampled_matrices,)
+
+
+###############################################################################
 
 
 @pytest.mark.parametrize("use_weight", [True, False])