+++ doc/whatsnew.rst
@@ -10,7 +10,9 @@ A catalog of new features, improvements, and bug-fixes in each release.
 v0.11.dev
 ---------
 
-- Enhance :func:`pyriemann.datasets.sample_gaussian_spd` adding support for dispersion defined as a covariance matrix.. :pr:`412` by :user:`thibaultdesurrel`
+- Enhance :func:`pyriemann.datasets.sample_gaussian_spd` adding support for dispersion defined as a covariance matrix. :pr:`412` by :user:`thibaultdesurrel`
+
+- Add :class:`pyriemann.clustering.GaussianMixture`. :pr:`411` by :user:`qbarthelemy`
 
 v0.10 (January 2026)
 --------------------